warning clean up

+++ cmd/smyrna/selection.c
@@ -1021,7 +1021,7 @@ int point_within_sphere_with_coords(float x0, float y0, float z0, float r,
 				    float x, float y,float z)
 {
 	float rr=(x-x0)*(x-x0)+(y-y0)*(y-y0)+(z-z0)*(z-z0);
-	rr=pow(rr,0.5);
+	rr=(float)pow(rr,0.5);
 	if (rr <= r)
 		return 1;
 	return 0;
